Add support for the Julia programming language
This adds support for coverage files from Julias package manager Pkg.jl.
The package manager outputs individual files for every source file in
the project, so the CoverageTools.jl package is used to combined them
all to the standardized lcov format. This lcov files is parsed from lua
into a table directly.

The parsing utility is added to the util module, since it should be
useful for other testframeworks and languages and not just Julia.