feat: support nested go modules (monorepo)
In case of a monorepo project (e.g. Python + Go) `go.mod` might be
located somewhere in nested directories, but not in the project root.

Usage of `Path:new('.')` won't work well in that case, because it is actually
a current working directory (cwd).

Change to `Path:new(vim.fn.expand("%:p"))` adds nested modules support,
since `go.mod` search will be started from the location of the current file
(opened in the buffer)

https://neovim.io/doc/user/builtin.html
> %		current file name
> :p		expand to full path