Add java jacoco coverage support
This is something I pulled together in a couple hours. Only tested with
jacocoTestReport.xml files generated by the gradle plugin. Added some
options to configure stuff.

Uses https://lunarmodules.github.io/luaexpat/ to parse XML. I tried to
use lua-xml reader but i could /not/ get it to compile on my computer.

I am a little shaky on some of the terminology but it works for me so i
thought i'd share.